+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## 0.2.3 (Unreleased)
 - Generics are now supported in Macros
+- Implement Digestible for `std::collections::{HashMap, HashSet, BTreeMap}`
 ## 0.2.2 (2023-10-13) 
 - Fixed Unresolved path for `core::any`